Avoid undefined behaviour.

+++ libraries/ZXSpectrum/Keyboard.c
@@ -44,7 +44,7 @@ int zxkeyset_for_port(uint16_t port, zxkeyset_t keystate)
 {
   int nzeroes;
 
-  nzeroes = my_clz(~port << 16);
+  nzeroes = my_clz((uint32_t) ~port << 16);
   return (~keystate >> (nzeroes * 5)) & 0x1F;
 }
